Control.Concurrent.STM.Connection

Documentation

data Connection r s

new

Arguments

:: Config
Queue size limits. Use defaultConfig for defaults.

-> Backend r s
Callbacks for accessing the underlying device.

-> IO (Connection r s)

Wrap a connection handle so sending and receiving can be done with STM transactions.

close :: Connection r s -> IO ()

Close the connection. After this, recv and send will fail.

close will block until all queued data has been sent and the connection has been closed. If close is interrupted, it will abandon any data still sitting in the send queue, and finish closing the connection in the background.

recv :: Connection r s -> STM (Maybe r)

Receive the next message from the connection. Return Nothing on EOF. Throw an exception if the Connection is closed, or if the underlying backendRecv failed.

This will block if no messages are available right now.

send :: Connection r s -> s -> STM ()

Send a message on the connection. Throw an exception if the Connection is closed, or if a previous backendSend failed.

This will block if the send queue is full.

cram :: Connection r s -> s -> STM ()

Like send, but never block, even if this causes the send queue to exceed configSendLimit.

isSendQueueEmpty :: Connection r s -> STM Bool

Test if the send queue is empty. This is useful when sending dummy messages to keep the connection alive, to avoid queuing such messages when the connection is already congested with messages to send.

bye :: Connection r s -> STM ()

Shut down the connection for sending, but keep the connection open so more data can be received. Subsequent calls to send and cram will fail.

Connection backend

data Backend r s

Connection I/O driver.

backendSend and backendRecv will often be called at the same time, so they must not block each other. However, backendRecv and backendSend are each called in their own thread, so it is safe to use IORefs to marshal state from call to call.

Constructors

Backend

Fields

backendRecv :: !(IO (Maybe r))

Receive the next message. Return Nothing on EOF.

backendSend :: !(Maybe s -> IO ())

Send (and flush) the given message.

If Nothing is given, shut down the underlying connection for sending, as backendSend will not be called again. If your device does not support this, do nothing.

backendClose :: !(IO ())

Close the device. backendSend and backendRecv are never called during or after this.

backendClose is called when the Connection is done using the device, even if you don't use close.

Configuration

data Config

Constructors

Config

Fields

configRecvLimit :: !(Maybe Int)

Default: Just 10

Number of messages that may sit in the receive queue before the Connection's receiving thread blocks. Having a limit is recommended in case the client produces messages faster than your program consumes them.

configSendLimit :: !(Maybe Int)

Default: Just 10

Number of messages that may sit in the send queue before send blocks. Having a limit is recommended if your program produces messages faster than they can be sent.

defaultConfig :: Config

Exceptions

data Error

Constructors

ErrorConnectionClosed String
connection closed

ErrorSentClose String
connection shut down for sending